+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ### is.codion.common.core
 - Observable added.
 - Item.get() renamed value(), no longer extends Supplier.
+- Observable no longer extends Supplier.
 ### is.codion.swing.common.model
 - TableSelectionModel, selectedIndex, selectedIndexes, selectedItem and selectedItems now Observable based.
 - TableSelectionModel.SelectedIndexes added.